codebase/requirements.txt
beautifulsoup4==4.11.2
boltons==23.0.0
boto3==1.18.33
more-itertools==9.1.0
numpy==1.21.6
openpyxl==3.0.9
optuna==3.2.0
pandas==1.3.5
pendulum==2.1.2
prefect==2.10.19
pretty_html_table==0.9.11
pydantic==1.10.11
PyNaCl==1.5.0
python-dotenv==0.19.0
pytest==7.3.1
python-dotenv
python-pptx==0.6.21
pytz==2021.1
redis==4.5.5
SQLAlchemy==1.4.36
xlrd==2.0.1
python-docx
python-pptx